// js/stat-evaluation-player/src/generated/BackboardPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Per-player accumulated backboard-bounce stats.
 */
export type BackboardPlayerStats = { count: number, is_last_backboard: boolean, last_backboard_time: number | null, last_backboard_frame: number | null, time_since_last_backboard: number | null, frames_since_last_backboard: number | null, };